requirements.txt
psycopg2==2.9.9
PyYAML==6.0.1
requests==2.32.3